fix(render): widen OT-range pointer chain to fix 64-bit truncation crash
InstDrawPerPlayer.otRangeNormal/otRangeSecondary stored a real native
OT-buffer pointer (from RenderBucket_AllocateOTRange, biased by a depth
offset) as plain `int`. RenderBucket_AddressSubOffset truncated the
pointer through (int)(u32)(uintptr_t)lhs to compute that bias, and every
consumer (RenderBucket_GetNormalOTEntry/GetClampedOTEntry and the whole
*AtRange dispatch family in RenderBucket_QueueExecute.c) reconstructed an
absolute pointer from the truncated int. On 64-bit hosts this discarded
the upper 32 bits, producing an invalid small pointer and crashing in
RenderBucket_LinkPrimRaw.

Same bug class as the prior InstDrawPerPlayer pointer-field fix, different
cluster: widened the whole chain to intptr_t (struct fields, the two
allocator/bias helper return types, ~16 *AtRange/*AtOTEntry function
signatures, and their local variable round-trips, including
MM_Title_SetTrophyDPP's e4/e8 locals). RenderBucket_AddressSub (a
same-buffer relative distance, not an absolute pointer) is correctly left
untouched -- truncating both operands there cancels out safely.

Confirmed via lldb the GPU 24-bit link-token bridge (native_gpu_links.c,
CtrGpu_PrimToOTLink24) is unrelated and was already correct; the bad
pointer never reached it.

Verified: macOS arm64 build clean (471 vs 473 pre-fix warnings), targeted
crash no longer reproduces across repeated runs. Binary now progresses
further and hits a separate, unrelated bug (a bogus TextureLayout pointer
in RenderBucket_DrawInstPrim_NormalAtOTEntry), tracked as a follow-up.

+++ game/230/MM_Title.c
@@ -291,8 +291,8 @@ void MM_Title_SetTrophyDPP(void)
 	struct InstDrawPerPlayer *idpp1;
 	struct InstDrawPerPlayer *idpp2;
 	struct Title *title = D230.titleObj;
-	int e4;
-	int e8;
+	intptr_t e4;
+	intptr_t e8;
 	int dc;
 
 	if (title == NULL)
